Create server.c
On the server side, you need to change the port number and make sure it matches that of the client side (PORT = XXXXX | line 12) as well as the IP address (CLIENT "XXXX.XXXX.XXX.XXX") which corresponds to the IP of your attacking machine. You can find it easily by performing an "ipconfig" under Windows or an "ifconfig" under Linux.